Feat : schedule 1,2,3 기능 완성

+++ src/main/java/com/green/backend/expert/service/ExpertService.java
@@ -29,8 +29,12 @@ public class ExpertService {
 
     // [1] 전문가 정보 등록
     public boolean createSpecialist(ExpertDTO expertDTO) {
-        try {expertRepository.save(expertDTO.toEntity());return true;
-        } catch (Exception e) {return false;}
+        try {
+            expertRepository.save(expertDTO.toEntity());
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
     }
 
     // [2] 전문가 목록 전체 조회
@@ -41,37 +45,39 @@ public class ExpertService {
     }
 
     // [3] 전문가 개별 조회
-    public ExpertDTO GetSpecialist(@PathVariable Long expertId) {
-        Expert expert =  expertRepository.findById(expertId)
+    public ExpertDTO GetSpecialist(Long expertId) {
+        Expert expert = expertRepository.findById(expertId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 전문가가 존재하지 않습니다."));
         return expert.toDto();
     }
 
     // [4] 전문가 수정
-    public boolean updateSpecialist(Long expertId, ExpertDTO expertDTO){
+    public boolean updateSpecialist(Long expertId, ExpertDTO expertDTO) {
         // 1. 수정할 대상이 있는지 먼저 찾기
         Expert expert = expertRepository.findById(expertId)
                 .orElse(null);
         if (expert == null) {
             return false;
         }
-            // 2. 찾아온 엔티티에 DTO 내용을 덮어씌우기 (Setter 사용)
-            expert.setExpertName(expertDTO.getExpertName());
-            expert.setExpertState(expertDTO.getExpertState());
-            expert.setExpertEmail(expertDTO.getExpertEmail());
-            expert.setExpertNumber(expertDTO.getExpertNumber());
-            expertRepository.save(expert);
-            return true;
-        }
+        // 2. 찾아온 엔티티에 DTO 내용을 덮어씌우기 (Setter 사용)
+        expert.setExpertName(expertDTO.getExpertName());
+        expert.setExpertState(expertDTO.getExpertState());
+        expert.setExpertEmail(expertDTO.getExpertEmail());
+        expert.setExpertNumber(expertDTO.getExpertNumber());
+        expertRepository.save(expert);
+        return true;
+    }
 
     // [5] 전문가 삭제
     // ExpertService.java
     public boolean deleteSpecialist(Long expertId) {
-        try { if(expertRepository.existsById(expertId)) {
+        try {
+            if (expertRepository.existsById(expertId)) {
                 expertRepository.deleteById(expertId);
-                return true;}
+                return true;
+            }
             return false; // 해당 ID가 없을 때
-        }catch (Exception e) {
+        } catch (Exception e) {
             return false;
         }
     }
